% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msy.R
\name{msy}
\alias{msy}
\title{Calculate maximum sustainable yield}
\usage{
msy(k, r)
}
\arguments{
\item{k}{k is the carrying capacity (individuals)}

\item{r}{r is the instrinsic growth rate (1/yr)}
}
\value{
Returns the maximum sustainable harvest for a fisher
}
\description{
Calculate maximum sustainable yield
}
\examples{
msy(300, 0.06)
}
