% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sushi_price.R
\name{sushi_price}
\alias{sushi_price}
\title{Calculate and print price of sushi}
\usage{
sushi_price(count, price)
}
\arguments{
\item{count}{Number of sushi orders}

\item{price}{Price of each order}
}
\description{
Calculate and print price of sushi
}
